+++ packages/flame/lib/src/nine_tile_box.dart
@@ -86,6 +86,6 @@ class NineTileBox {
     final xSrc = sprite.src.left + _tileSizeDouble * i;
     final ySrc = sprite.src.top + _tileSizeDouble * j;
     final src = Rect.fromLTWH(xSrc, ySrc, _tileSizeDouble, _tileSizeDouble);
-    c.drawImageRect(sprite.image, src, dest, BasicPalette.white.paint);
+    c.drawImageRect(sprite.image, src, dest, BasicPalette.white.paint());
   }
 }