+++ packages/flame/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [next]
  - Fix input bug with other anchors than center
  - Moving tutorials to the Flame main repository
+ - Transforming PaletteEntry.paint to be a method instead of a getter
+ - Adding some more basic colors entries to the BasicPalette
 
 ## 1.0.0-rc8
  - Migrate to null safety